sched/fair: correct fastpath
idle_get_state_idx() returns the idle state of the cpu - 0 being the
shallowest and higher values indicate deeper idle state.

The code currently chooses prev_cpu for a idle state of less than 1. 1
on our platforms corresponds to a deep idle state,and for mid and high
capacity cpus exiting from it incurs additional latency.

So amend the code such that for mid and high cluster cpus we limit
prev_cpu fastpath to the shallowest idle state. For min capacity cpus
using idle state of 1 is fine, as it doesn't incur huge idle exit
latencies.

Change-Id: Idb544015c8b533a2b1ad9c1355db3e975162f074
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>